+++ lib/src/node/legacy/value/list.dart
@@ -4,10 +4,8 @@
 
 import 'package:js/js.dart';
 
-import 'dart:js_util';
-
 import '../../../value.dart';
-import '../../utils.dart';
+import '../../reflection.dart';
 import '../value.dart';
 
 @JS()
@@ -18,31 +16,31 @@ class _NodeSassList {
 
 /// Creates a new `sass.types.List` object wrapping [value].
 Object newNodeSassList(SassList value) =>
-    callConstructor(listConstructor, [null, null, value]) as Object;
+    legacyListClass.construct([null, null, value]);
 
-/// The JS constructor for the `sass.types.List` class.
-final Function listConstructor = createClass('SassList',
+/// The JS `sass.types.List` class.
+final JSClass legacyListClass = createJSClass('sass.types.List',
     (_NodeSassList thisArg, int? length,
         [bool? commaSeparator, SassList? dartValue]) {
   thisArg.dartValue = dartValue ??
       // Either [dartValue] or [length] must be passed.
       SassList(Iterable.generate(length!, (_) => sassNull),
           (commaSeparator ?? true) ? ListSeparator.comma : ListSeparator.space);
-}, {
-  'getValue': (_NodeSassList thisArg, int index) =>
-      wrapValue(thisArg.dartValue.asList[index]),
-  'setValue': (_NodeSassList thisArg, int index, Object value) {
-    var mutable = thisArg.dartValue.asList.toList();
-    mutable[index] = unwrapValue(value);
-    thisArg.dartValue = thisArg.dartValue.withListContents(mutable);
-  },
-  'getSeparator': (_NodeSassList thisArg) =>
-      thisArg.dartValue.separator == ListSeparator.comma,
-  'setSeparator': (_NodeSassList thisArg, bool isComma) {
-    thisArg.dartValue = SassList(thisArg.dartValue.asList,
-        isComma ? ListSeparator.comma : ListSeparator.space,
-        brackets: thisArg.dartValue.hasBrackets);
-  },
-  'getLength': (_NodeSassList thisArg) => thisArg.dartValue.asList.length,
-  'toString': (_NodeSassList thisArg) => thisArg.dartValue.toString()
-});
+})
+  ..defineMethods({
+    'getValue': (_NodeSassList thisArg, int index) =>
+        wrapValue(thisArg.dartValue.asList[index]),
+    'setValue': (_NodeSassList thisArg, int index, Object value) {
+      var mutable = thisArg.dartValue.asList.toList();
+      mutable[index] = unwrapValue(value);
+      thisArg.dartValue = thisArg.dartValue.withListContents(mutable);
+    },
+    'getSeparator': (_NodeSassList thisArg) =>
+        thisArg.dartValue.separator == ListSeparator.comma,
+    'setSeparator': (_NodeSassList thisArg, bool isComma) {
+      thisArg.dartValue = SassList(thisArg.dartValue.asList,
+          isComma ? ListSeparator.comma : ListSeparator.space,
+          brackets: thisArg.dartValue.hasBrackets);
+    },
+    'getLength': (_NodeSassList thisArg) => thisArg.dartValue.asList.length
+  });