+++ test/legacy_node_api/value/boolean_test.dart
@@ -25,11 +25,6 @@ void main() {
       expect(value, isJSInstanceOf(sass.types.Boolean));
       expect(value.getValue(), isFalse);
     });
-
-    test("has a useful .constructor.name", () {
-      expect(parseValue<NodeSassBoolean>("true").constructor.name,
-          equals("SassBoolean"));
-    });
   });
 
   group("from a constant", () {
@@ -43,10 +38,6 @@ void main() {
       var value = sass.types.Boolean.FALSE;
       expect(value, isJSInstanceOf(sass.types.Boolean));
       expect(value.getValue(), isFalse);
-    });
-
-    test("has a useful .constructor.name", () {
-      expect(sass.types.Boolean.FALSE.constructor.name, equals("SassBoolean"));
     });
   });