+++ tool/grind.dart
@@ -30,7 +30,9 @@ void main(List<String> args) {
   pkg.homebrewFormula.value = "sass.rb";
   pkg.jsRequires.value = [
     pkg.JSRequire("chokidar", target: pkg.JSRequireTarget.cli),
-    pkg.JSRequire("readline", target: pkg.JSRequireTarget.cli)
+    pkg.JSRequire("readline", target: pkg.JSRequireTarget.cli),
+    pkg.JSRequire("immutable", target: pkg.JSRequireTarget.all),
+    pkg.JSRequire("util", target: pkg.JSRequireTarget.all),
   ];
   pkg.jsModuleMainLibrary.value = "lib/src/node.dart";
   pkg.jsDevFlags.value.add("-Dnew-js-api=true");